services/transaction-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
psycopg2-binary>=2.9
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
aio-pika>=9.4.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.6
